==> Code/01_Data_Gathering.ipynb <==
# Data Gathering

import pandas as pd
import numpy as np
import requests
import time
import warnings; warnings.simplefilter('ignore')

###### These are the URLS of the sub reddits that I will be gathering my data from
Because Email marketing is newer sub reddit compared to the others, i decided to gather data from the practically identical subreddit, email, and combine the two

url1 = 'https://www.reddit.com/r/PPC.json'
url2 = 'https://www.reddit.com/r/SEO.json'
url3 = 'https://www.reddit.com/r/Emailmarketing.json'
url4 = 'https://www.reddit.com/r/email.json'

#### This for loop gathers reddit data 25 posts at a time, and goes through the following steps
 1. Requests to gather the data from reddit, and if the status code does not equal 200, breaks the loop
 2. Turns the json file that is pull from reddit, and turns it into a dictionary, which is easier to work with in python
 3. The loop then pulls the relevant data from the dictionary, and adds it to the list "posts"
 4. At the end of the for loop, it pulls the URL of the 26th post adds it to the current url, and restarts the loop, pulling the next 25 posts
 5. The results are then saved as a CSV file
 6. This for loop is repeated for each url

url_dict = {url1:'PPC',url2:"SEO",url3:"email1",url4:"email2"}

for url in url_dict:
    posts = []
    after = None
    for i in range(40):
        if after == None:
            current_url = url
        else:
            current_url = url + '?after=' + after
        print(current_url)
        res = requests.get(current_url, headers ={'User-agent': 'Matthew Parker'})
        if res.status_code != 200:
            print('Status Error', res.status_code)
            break
        current_dict= res.json()
        current_posts = [p['data'] for p in current_dict['data']['children']]
        posts.extend(current_posts)
        after = current_dict['data']['after']
        time.sleep(2)
    pd.DataFrame(posts).to_csv('../Data/'+url_dict[url]+'.csv')
