==> Code/04_Modeling_Multinomial_Naive_Bayes.ipynb <==
# Mulitnomial Naive Bayes

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.grid_search import GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import TfidfVectorizer 
import pickle
import warnings; warnings.simplefilter('ignore')
np.random.RandomState(42)

X_train = pd.read_csv("../Data/X_train.csv",index_col=0)
y_train = pd.read_csv("../Data/y_train.csv",index_col=0)
with open('../Assets/custom_stop_words.pkl','rb') as f:
    custom_stop_words = pickle.load(f)

This cell instantiates a Pipeline, which will process the data, and  determines which model I will use in this notebook, which is the Multinomial Naive Bayes model.

pipe = Pipeline([
    ("count", TfidfVectorizer()),
    ("MultiNB", MultinomialNB())
])

This dictionary lists the hyper parameters that I want my Gridsearch to test, which will then pick a single model. Gridsearch will then use the model that had the best accuracy.

params = {
    "count__stop_words":[custom_stop_words],
    "count__min_df":[5],
    'MultiNB__alpha': [1,.5,.1]
}

The alpha value in the Multinomial Naive Bayes model dictates how strong the regularization is. The min_df parameter is saying that in order for a word to be considered in the model, it must show up in at least 5 documents. This is done in order to reduce the overall number of feature. Most is not all of the words being eliminated should have no importance in determining each class if it shows up in less than 5 posts.

gs_MultiNB = GridSearchCV(pipe,param_grid=params)

gs_MultiNB.fit(X_train['Total_text'],y_train['subreddit'])

gs_MultiNB.best_params_['MultiNB__alpha']

The alpha that performed the best in the model was .5

###### Saving our  Multinomial Naive Bayes Model to be evaluated later

with open('../Assets/MultiNB.pkl','wb+') as f:
    pickle.dump(gs_MultiNB,f)
